Convert Views to make edit work
