+++ iOS/Financer/Financer/Views/Components/LegalPerson/LegalPersonEditor.swift
@@ -1,0 +1,20 @@
+//
+//  LegalPersonEditor.swift
+//  Financer
+//
+//  Created by Julian Schumacher on 01.01.23.
+//
+
+import SwiftUI
+
+struct LegalPersonEditor: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct LegalPersonEditor_Previews: PreviewProvider {
+    static var previews: some View {
+        LegalPersonEditor()
+    }
+}
