+++ iOS/Financer/Financer/Views/LegalPerson/EditLegalPerson.swift
@@ -1,0 +1,20 @@
+//
+//  EditLegalPerson.swift
+//  Financer
+//
+//  Created by Julian Schumacher on 01.01.23.
+//
+
+import SwiftUI
+
+struct EditLegalPerson: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct EditLegalPerson_Previews: PreviewProvider {
+    static var previews: some View {
+        EditLegalPerson()
+    }
+}
